pulpo_project/requirements.txt
django
django-extensions
djangorestframework
django-auth-ldap
fontawesomefree
PyMoe
python-dotenv
gunicorn
requests
wagtail
psycopg2-binary
django-allauth